Refactor apply pruning proof functions
